crud hero

+++ ldvelh/src/main/java/com/nofinish/ldvelh/repo/BookRepo.java
@@ -3,7 +3,7 @@ package com.nofinish.ldvelh.repo;
 import com.nofinish.ldvelh.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepo extends JpaRepository<Book, Integer> {
+public interface BookRepo extends JpaRepository<Book, Long> {
 
     Book findByTitle(String title);
